Added ability defines for all abilities up to gen 8, and changed all Pokemon abilities to match gen 8

+++ include/battle.h
@@ -655,7 +655,7 @@ extern u16 gChosenMove;
 extern u16 gCalledMove;
 extern u8 gCritMultiplier;
 extern u16 gBattleWeather;
-extern u8 gLastUsedAbility;
+extern u16 gLastUsedAbility;
 extern u8 gBattlerInMenuId;
 extern u8 gPotentialItemEffectBattler;
 extern u8 gBattlersCount;
